fixed problem with frequency estimation when frequency shifts are very rapid. in this example, frequencies went from 0 to 1 very quickly resulting in a
000001000001000000011111111111111111111111111111111111111111111111111

pattern of observations. To speed up freq estimation, we had restricted
the interval used to 6 pivots after the last 0. With stiff prior on frequency
change, the frequency had not reached 1 by that point.

to fix this, a few oversights in argument passing got fixed, frequency estimation
made less stiff, and the number of pivots after the last 0 increased to 12