Start on continuous prediction restructure.
This swaps seasons for pairs of initial and final timepoints. We want
the set of predictors and frequencies for every tip for each initial
timepoint and to use these to project forward to the final timepoints.
These need not be 1 year apart. So far, have basic setup, need to
standardize predictors and do the actual fitting.