Revise tolerance predictors and include HA1 epitope mask.
fitness_predictors now runs the code in fitness_tolerance to assign
tol and tol_ne predictors to nodes. I messed with the epitope mask
code in fitness_tolerance, because epitope_mask seemed to be a different
form than the previous code was expecting. I believe it's working as
intended given testing in auspice. I also updated the H1/H3 alignment
to align WSN to Beijing/92.