Fix bug in which strains were being matched between FASTA, HI and source
I discovered this with noticing that "B/PHUKET/3073/2013" had no matches in live site. This was because the "reference" strain (with sequence in ref.json) had the name PHUKET, while the HI titer file listed this as "B/Phuket/3073/2013". Resolving this across strains has resulted in many hundreds of additional measurements being brought into the live build.