auspice: Update serum id creation
Replace any characters in the serum name that are not alphanumeric,
dashes or underscores with an underscore to create a serum id that can
used as a query selector. I know that replacing with just an underscore
is not the most robust way to create unique IDs, but this is just a
quick fix to get the private flu site working again.

Also refactored the three separate serum id creations into a single
function so that I don't have to change the code in three places.